requirements.txt
streamlit
torch
openai-whisper
TTS
soundfile
numpy
